+++ benchmark/indexes/jaccard/plot_threshold_benchmark.py
