+++ benchmark/sketches/similarity_benchmark.py
